fix creature difficulty, update spawn masks, fix some startup errors, update TODO.md

+++ src/server/game/Entities/Pet/Pet.cpp
@@ -1391,8 +1391,8 @@ void Pet::_SaveAuras(SQLTransaction& trans)
         stmt->setUInt32(index++, m_charmInfo->GetPetNumber());
         stmt->setUInt64(index++, casterGUID);
         stmt->setUInt32(index++, itr->second->GetId());
-        stmt->setUInt8(index++, effMask);
-        stmt->setUInt8(index++, recalculateMask);
+        stmt->setUInt32(index++, effMask);
+        stmt->setUInt32(index++, recalculateMask);
         stmt->setUInt8(index++, itr->second->GetStackAmount());
         stmt->setInt32(index++, damage[0]);
         stmt->setInt32(index++, damage[1]);
